Design: 회원가입 완료 페이지 center에 모듈`처럼` 뜨게끔 디자인 수정
